Dashboard, NearbyShop, NearbyShops, PreferredShop, PreferredShops components
Add routes to Dashboard, NearbyShops, PreferredShops components

Create a dashboard component to present the application

Set up the NearbyShops component to display multiple NearbyShop components

Set up the PreferredShops component to display multiple PreferredShop components

Modify the Navbar component to display the login and logout icon

Add to attribute

Add to attribute

Set up the NearbyShop component to display the title, a random image, the like and dislike button, and the distance.

Set up the PreferredShop component to display the title, a random image, the remove button, and the distance.